+++ src/main/docker/entrypoint.sh
@@ -12,13 +12,14 @@ DEBUG_JAVASCRIPT=${DEBUG_JAVASCRIPT:-false}
 DOCKER_HOST=${DOCKER_HOST:-}
 FASTER_CODEC=${FASTER_CODEC:-false}
 GOOGLE_API_KEY=${GOOGLE_API_KEY:-}
-GPS_TOPICS='['`echo ${GPS_TOPICS} | perl -pe 's#([/\w+]+)#"\1"#g'`']'
+GPS_TOPICS='['$(echo "${GPS_TOPICS}" | perl -pe 's#([/\w+]+)#"\1"#g')']'
 LDAP_BINDDN=${LDAP_BINDDN:-}
 LDAP_BIND_PASSWORD=${LDAP_BIND_PASSWORD:-}
 LDAP_SEARCH_BASE=${LDAP_SEARCH_BASE:-}
 LDAP_SERVER=${LDAP_SERVER:-}
 LDAP_USER_PATTERN=${LDAP_USER_PATTERN:-}
-METADATA_TOPICS='['`echo ${METADATA_TOPICS} | perl -pe 's#([/\w+]+)#"\1"#g'`']'
+METADATA_TOPICS='['$(echo "${METADATA_TOPICS}" | perl -pe 's#([/\w+]+)#"\1"#g')']'
+OPEN_WITH_URLS=${OPEN_WITH_URLS:-"{'Webviz':['https://webviz.io/app/?', 'remote-bag-url']}"}
 SCRIPT_TMP_PATH=${SCRIPT_TMP_PATH:-/scripts}
 TILE_HEIGHT_PX=${TILE_HEIGHT_PX-256}
 TILE_MAP_URL=${TILE_MAP_URL-"http://{a-d}.tile.stamen.com/terrain/{z}/{x}/{y}.jpg"}
@@ -26,7 +27,7 @@ TILE_WIDTH_PX=${TILE_WIDTH_PX-256}
 USE_BING=${USE_BING:-false}
 USE_MAPQUEST=${USE_MAPQUEST:-true}
 USE_TILE_MAP=${USE_TILE_MAP:-true}
-VEHICLE_NAME_TOPICS='['`echo ${VEHICLE_NAME_TOPICS} | perl -pe 's#([/\w+]+)#"\1"#g'`']'
+VEHICLE_NAME_TOPICS='['$(echo "${VEHICLE_NAME_TOPICS}" | perl -pe 's#([/\w+]+)#"\1"#g')']'
 
 # Don't overwrite an existing settings.yml file, but if we don't have one, write all of
 # the variables out to it
@@ -51,6 +52,7 @@ ldapSearchBase: ${LDAP_SEARCH_BASE}
 ldapServer: ${LDAP_SERVER}
 ldapUserPattern: ${LDAP_USER_PATTERN}
 metadataTopics: ${METADATA_TOPICS}
+openWithUrls: ${OPEN_WITH_URLS}
 scriptTmpPath: ${SCRIPT_TMP_PATH}
 tileHeightPx: ${TILE_HEIGHT_PX}
 tileMapUrl: ${TILE_MAP_URL}